QAS/requirements.txt
pytest==6.2.5
pytest-selenium==4.0.0
webdriver-manager==3.8.5
requests==2.28.2
selenium-requests==2.0.3